Trigger replication on putACL only for CRR
When ACL are updated, check the updated replication info: if replicating
to cloud locations (i.e. zenko's multi-backend replication), then we
do not update replication info, to ensure ACL are not replicated.

However, if we are replicating to CRR location, then we should replicate
ACL, and we will now reset the replicationInfo.

Issue: CLDSRV-638