Remove useless storage class update
Code tried to set storage-class from bucket's default location when
cold, but
- a bucket cannot have a cold location as default;
- storage location is per-object, not per bucket;
- it is set as expected in `collectResponseHeaders()`, in the next line.

Take the chance to create proper "mocks" for the different "states" of
archived/cold objects. Eventually this should be moved to Arsenal,
along with models, to create a corpus of possible object states and
enforce valid transitions in the model APIs.

Issue: CLDSRV-638