Allow configuring valid storage classes
The only supported storage class and should be STANDARD. We have no use
for other storage classes (in requests), and should not try to accept
anything else as it may affect the understanding apps have of their
use of the storage.

However, to ease transition, allow overriding the list of "valid"
storage classes by specifying the `VALID_STORAGE_CLASSES` environment
variable. This is a comma-separate list of supported values.

Practically, the parameter is not very useful:
* It is _validated_ in putObject, but not otherwise used: if storage
 class is specified, it has not effect on metadata or storage;
* It is added to the object when calling `copyObject`, with no further
  validation. To match behavior of 7.x branches, all used values should
  be added to the env variable;
* In initiate MPU, the storage class is stored in the MPU overview doc
  only it is also one of 'standard', 'standard_ia', 'reduced_redundancy'
  (in lowercase!).

By setting `VALID_STORAGE_CLASSES` appropriately, it is possible to get
the same behavior as 7.x branch; and this will be cleaned'up in 9.1
branch through the CLDSRV-639 ticket.

Issue: CLDSRV-638