Support batchDelete route when bucketName is not present
This parameter was not originally present, and is not enforced at the
moment (even though it seems to be set on every "production" call from
backbeat). At the moment, this parameter is mostly used only for quotas'
inflight update.

This should be cleaned' in CLDSRV-643, to settle on a proper API.
However, gracefully handle the case where the field is not set, so that:
- We ensure no regression in S3C when upgrading from 7 to 9;
- We give a change to multibackend's routeBackbeat tests to pass
  (CLDSRV-394) as they are still using the old route without bucket.

Issue: CLDSRV-638