Fix deletion of wrong data in abort MPU cond. scenarios
Add functional tests to cover scenarios where aborting an MPU
could delete data from a different, completed MPU with the
same object key.

This prevents a regression of a bug where the wrong data was
deleted because the uploadId was not being checked when
looking for object metadata to clean up during an abort.

Issue: CLDSRV-668