CLDSRV-734: Revert back from go to python supervisord
This will fix fluent-bit access to s3 log files

The go version is not a perfect replacement it has multiple changes that would require to update supervisor conf.

Breaking changes includes:

- log files name changes
  - process number starts count at 1 instead of 0
  - include full process_name with number in logfile
  - s3_1-1.log instead of s3-0.log
- no supervisorctl, replaced by superivsord ctl with less option
- no interactive supervisor ctl, if used without command it rewrite the superivsord.sock and then we lose access to supervisord ctl
- replaced priority with depends_on
- replaced %(ENV_X)s with envFiles

Overrall we’d like to avoid breaking changes until we resolve correct usage of the go version and apply it in every component at once.