Dummyservice adaptation due to aws-sdk migration
This commit aims to add the necessary changes to the DummyService.js
file to ensure compatibility with the latest version of the AWS SDK.
The modifications include updating method calls, handling responses,
and ensuring that all interactions with AWS services are compliant
with the new SDK standards. I purposefully added the support for both
callbacks and promises to avoid breaking changes in the codebase.

Issue: CLDSRV-724