Serve capabilities from config if present
The capabilities returned by getCapabilities API will not be directly
set in config by ZKOP, so that they are simply _served_ by Cloudserver.

Old fields are keps for backwards compatibility, but the goal is really
that ZKOP is free to put _any_ content there, as needed by pensieve-api:
this will in particular help to minimize the changes required when
adding a new kind of location.

Issue: CLDSRV-599