checking out on latest commit on the pykmip repo
The image build was failing because we were trying to install pykmip
with python v2 (setup.py install uses python command which points to python v2
by default in alpine image). This change checks out the latest commit and
installs pykmip with pip3 as this is the one now supported by pykmip.
Issue: CLDSRV-806